Info endpoint is now unsecured, and a stub for secureinfo endpoint has been added. Added in some basic documentation about the info endpoint, and also corrected an issue noticed with the metrics documentation that may be attributable to a copy paste.
