+++ documentation/api/README.md
@@ -12,5 +12,6 @@ The following are the root routes on the HTTP API surface.
 | [`/trace`](trace.md) | Captures traces of processes without using a profiler. |
 | [`/metrics`](metrics.md) | Captures metrics of a process. |
 | [`/logs`](logs.md) | Captures logs of processes. |
+| [`/info`](info.md) | Gets info about Dotnet Monitor. |
 
 The `dotnet monitor` tool is able to detect .NET Core 3.1 and .NET 5+ applications. When connecting to a .NET Core 3.1 application, some information may not be available and is called out in the documentation.
