+++ documentation/api/metrics.md
@@ -33,8 +33,6 @@ Authentication is not enforced for this route.
 |---|---|---|---|
 | 200 OK | | A list of metrics for a single process in the Prometheus exposition format. | `text/plain` |
 | 400 Bad Request | [ValidationProblemDetails](definitions.md#ValidationProblemDetails) | An error occurred due to invalid input. The response body describes the specific problem(s). | `application/problem+json` |
-| 401 Unauthorized | | Authentication is required to complete the request. See [Authentication](./../authentication.md) for further information. | |
-| 429 Too Many Requests | | There are too many trace requests at this time. Try to request a trace at a later time. | |
 
 ## Examples
