+++ documentation/openapi.json
@@ -754,19 +754,43 @@
         }
       }
     },
-    "/info": {
+    "/secureinfo": {
       "get": {
         "tags": [
           "Diag"
         ],
-        "summary": "Gets the Dotnet Monitor Version (other configuration information may be added in the future).",
-        "operationId": "GetInfo",
+        "summary": "Gets information about Dotnet-Monitor that should be secured behind authentication (currently a stub).",
+        "operationId": "GetSecureInfo",
         "responses": {
           "400": {
             "$ref": "#/components/responses/BadRequestResponse"
           },
           "401": {
             "$ref": "#/components/responses/UnauthorizedResponse"
+          },
+          "200": {
+            "description": "Success",
+            "content": {
+              "application/json": {
+                "schema": {
+                  "$ref": "#/components/schemas/DotnetMonitorSecureInfo"
+                }
+              }
+            }
+          }
+        }
+      }
+    },
+    "/info": {
+      "get": {
+        "tags": [
+          "Info"
+        ],
+        "summary": "Gets the Dotnet Monitor Version (other configuration information may be added in the future).",
+        "operationId": "GetInfo",
+        "responses": {
+          "400": {
+            "$ref": "#/components/responses/BadRequestResponse"
           },
           "200": {
             "description": "Success",
@@ -1011,6 +1035,10 @@
             "description": "Set to true to collect logs at the application-defined categories and levels."
           }
         },
+        "additionalProperties": false
+      },
+      "DotnetMonitorSecureInfo": {
+        "type": "object",
         "additionalProperties": false
       },
       "DotnetMonitorInfo": {
